requirements.txt
flask==3.1.2
flask-cors==6.0.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.181.0
openai==1.107.1
python-dotenv==1.1.1
